functions/solutions/train_model/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
db-dtypes
scikit-learn
joblib